Rename `markets` to `market`
A module name does not need to be plural.
